Added ability to batch-fetch secrets
Since you may want multiple secrets fetched at a atime, this PR adds
this capability to the codebase. The API is a bit weird here but we try
to accomodate it and output JSON just like the API does (which probably
should be able to be specified as the output format.