Fix handling of defaulted account values in Client
Old code was effectively hardcoded to always use "default" as the
account name when an account name was not provided, including times
where this data was supposed to be pulled from .conjurrc. This change
ensures that we only default the account name iff it's the only filed
missing from a valid configuration.