Add UTs and integration test and test artifacts
This commits adds new tests and their test artifacts. As a new requirement, we are removing the ability to pass in configuration details via the commandline. Therefore, this commit also removes the relevant tests and updates the infrastructure to accept configuration details from a generated conjurrc file. For backwards capability purposes though, Client will be untouched.