Add init CLI command logic
This PR adds the logic for the init CLI command and respective testing. We got feedback from the field that users find it difficult to remember their accounts when initializing the CLI so we added the ability to discover the account for them to avoid prompting them for one. As a new requirement, we are removing the ability to pass in configuration details via the commandline. Therefore, this PR also removes the relevant tests and updates the testing infrastructure to accept configuration details from a generated conjurrc file. For backwards capability purposes though, Client will be untouched.