Add host and user implementation and integration tests (#148)
This PR adds the logic and integration tests for host/user commands (user- rotate-api-key, change-password and host- rotate-api-key). This PR also fixes a bug that was discovered when logging in as another user when already logged into the CLI. When a logged in attempted to login as another user, an extra few characters were being written to the bottom of the netrc. This broke the netrc format and resulted in and error for all follow-up commands. The problem originated in file permissions r+ instead of w.