Add CLI setup script for macOS
Previously the CLI on macOS environments was taking upwards of 15+ seconds on each run. This is because of a security policy on macOS machines that extracts executables and scans them. To account for this, we will pack the CLI as a directory instead of a single executable file and create a symbolic link between the directory and /usr/local/bin. This script hides away these details for the user