Remove publishing to PyPI (see description for more info)
We push the CLI and SDK as a single package to PyPi so that users with Python installed can easily pull the CLI for human-use and SDK for application-use. The code for the SDK and CLI are together under a single repo and pushed to PyPi as a single library. Because only the CLI is to be promoted to GA (v7.0.0) we cannot continue to push both the CLI and the SDK together to PyPI and there is no way (at the moment) to separate the artifacts into two different libraries so that the client can remain < v1.0.0.
An [issue](https://github.com/cyberark/conjur-api-python3/issues/211) was open to properly address this.