Update change-password tests to change non-admin passwords and fix failing UTs (#220)
Password complexity restriction was a feature introduced in a relatively newer version of Conjur. Before release we need to check against these older server versions and therefore when running the integration tests for change-password, the tests were failing and the actual password was being changed which leads to failing tests. This fixes this by attempting to change the password on a dummy user.

This PR also gives our broken UT build. The instability was coming the way at which the Jenkins pipeline is running the tests.