Add error handling for when conjurrc configuration is missing (#226)
Previously, when the conjurrc was lacking mandatory parameters, an unclear error was being returned. The error was just returning the name of the parameter missing, for example 'cert_file' without really explaining the problem or how to solve it. This becomes more apparently and visible because the conjurrc parameter naming was changed. This commit adds the necessarily error handling so that the user will understand how to proceed