Save credentials to system keystore (#233)
Adds the ability to add credentials to a system's keyring. This is the default flow. If a supported keyring cannot be found then we will resort to writing credentials to the netrc with a warning to logout after use. This PR does not include tests because this requires a change in infra which will lead to a very big PR. This is the first PR in a series of multiple PRs.