Enable Pypi push
Previously we decided that we were not going to push to Pypi because only the CLI is being promoted to GA and the pushed package includes both the CLI and SDK. Now, we changed this decision and will now push to Pypi and detail that the SDK will remain a Community-level project so that all will continue to benefit from the project's developments. This commit also adds entries to the CHANGELOG to detail all breaking changes.