Remove irrelevant list call
Previously, we had an 'if/else' statement where the 'else' was never being reached. It was originally added to differentiate when the request had constraints and when it didn't. Being that in the end, it doesn't make a difference and requests are correctly invoked, this should be removed as it is never being reached.