Publish to PyPi with manual condition
We do not want the package to be published to PyPi before the release
is official. Thus, we can't have just a trigger on the tag because the
tag is created as part of the release process and not when it is done.

To overcome this, we can add a build param that can be run after the
release is official.